jarvis-ar-visual-system/jarvis-ar-visual-system/requirements.txt
numpy
opencv-python
tensorflow
torch
scikit-learn
matplotlib
Pillow
pyglet
pyqt5
mlflow
wandb
pyyaml